Further enhancements to the application: - The amdModules now only loads the modules used in the application. In a continuous build, the amd modules would be an ever growing set that would not handle removed modules - The amd config script is only loaded once - An individual amd start script is created for each index.html file when not writing inline.
