+++ lib/start-template.txt
@@ -1,0 +1,26 @@
+require([
+  <%= names %>
+], function (<%= objects %>) {
+  var adoptables = [<%= adoptables %>];
+  var isVendor = new RegExp('<%= vendor %>(.*js)');
+  function recursiveRequire(i, scripts){
+    if (i >= scripts.length) {
+      return;
+    }
+    var scriptToLoad = scripts[i];
+    if (scriptToLoad.src) {
+      require([scriptToLoad.src], function() {
+        if (isVendor.test(scriptToLoad.src)) {
+          adoptables.forEach(function(adoptable){
+            enifed(adoptable.name, [], function(){return adoptable.obj;});
+          });
+        }
+        recursiveRequire(++i, scripts);
+      });
+    } else {
+      eval(scriptToLoad.code);
+      recursiveRequire(++i, scripts);
+    }
+  }
+  recursiveRequire(0, [<%= scripts %>]);
+});
