+++ tests/dummy/app/components/esri-map.js
@@ -1,7 +1,6 @@
 import Component from '@ember/component';
 
 import Map from 'esri/map';
-import Graphic from 'esri/graphic';
 
 const EsriMapComponent = Component.extend({
     didInsertElement() {
