+++ tests/dummy/app/index.html
@@ -17,9 +17,11 @@
   <body>
     {{content-for "body"}}
 
+    <script>console.log('Should would before vendor load');</script>
     <script src="{{rootURL}}assets/vendor.js"></script>
-    {{content-for "amd-modules"}}
+    <script>console.log('Should work after vender load but before app load');</script>
     <script src="{{rootURL}}assets/dummy.js"></script>
+    <script>console.log('Should work after app load');</script>
 
     {{content-for "body-footer"}}
   </body>
