+++ tests/index.html
@@ -20,11 +20,10 @@
   <body>
     {{content-for "body"}}
     {{content-for "test-body"}}
-    
+
 
     <script src="/testem.js" integrity=""></script>
     <script src="{{rootURL}}assets/vendor.js"></script>
-    {{content-for "amd-modules"}}
     <script src="{{rootURL}}assets/test-support.js"></script>
     <script src="{{rootURL}}assets/dummy.js"></script>
     <script src="{{rootURL}}assets/tests.js"></script>
